% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_number.R
\name{clean_number}
\alias{clean_number}
\title{Clean nunber. This function remove NA data and flagged
data from us hcn files. Remove all the flags in the
raw data and preserve only numeric data}
\usage{
clean_number(x)
}
\arguments{
\item{x}{A string with the ushcn value to be cleaned}
}
\value{
A double value with the us hcn data cleaned
}
\description{
Clean nunber. This function remove NA data and flagged
data from us hcn files. Remove all the flags in the
raw data and preserve only numeric data
}
\keyword{internal}
