% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chelsa_files.R
\name{get_chelsa_files}
\alias{get_chelsa_files}
\title{Get Chelsa time series files}
\usage{
get_chelsa_files(variable, model, scenario, yearStart, yearEnd)
}
\arguments{
\item{variable}{A string with the possible variable. Could be pr, tas, tasmax or tasmin}

\item{model}{A string with the CMIP5 model ACCESS1-3, CESM1-BGC, CMCC-CM or MIROC5}

\item{scenario}{A string with CMIP5 scenario, historical, rcp45 or rcp85}

\item{yearStart}{The initial year in the CHELSA raster file}

\item{yearEnd}{The final year in the CHELSA raster file}
}
\value{
Download Chelsa required files
}
\description{
Get Chelsa time series files
}
\examples{
get_chelsa_files("test")
}
