% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ushcn_data.R
\name{get_ushcn_data}
\alias{get_ushcn_data}
\title{Get ushcn data stored from tar.gz files for an specific variable}
\usage{
get_ushcn_data(variable)
}
\arguments{
\item{variable}{String, could choose between, tmax, tmin, tavg, prcp. Alias
for prcp (pr) and tavg (tas) are avaliable to be consistent with chelsa data.}
}
\value{
A time serires data frame with station, year, and month from 1 to 12
}
\description{
Get ushcn data stored from tar.gz files for an specific variable
}
\examples{
get_ushcn_data("tmin")
}
