% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ushcn_files.R
\name{get_ushcn_files}
\alias{get_ushcn_files}
\title{Get US HCN files}
\usage{
get_ushcn_files(variable)
}
\arguments{
\item{variable}{A string with the climatic variable avaliable in US hcn monthly
database. Possible options are prcp, tavg, tmin, tmax, and test.
If test is provided checks if US HCN server is online}
}
\value{
Download us hcn required files
}
\description{
Get US HCN files
}
\examples{
get_ushcn_files("test")
}
