% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ushcn_files.R
\name{read_ushcn_files}
\alias{read_ushcn_files}
\title{Read a station text file from ushcn downloaded tar.gz file.
The US HSN data is stored in a .tar.gz file with all the stations.
This function read each file stored in the .tar.gz once
by time. The first file of the .tar.gz correspond
to the directory containing all the data,}
\usage{
read_ushcn_files(file, stationIndex)
}
\arguments{
\item{file}{the tar.gz ushcn file downloaded}

\item{stationIndex}{Index of the texfile to be reader as data.frame}
}
\value{
A tibble with the ushcn data
}
\description{
Read a station text file from ushcn downloaded tar.gz file.
The US HSN data is stored in a .tar.gz file with all the stations.
This function read each file stored in the .tar.gz once
by time. The first file of the .tar.gz correspond
to the directory containing all the data,
}
\keyword{internal}
